status bar and router use theme
